+++ flutter_google_places_sdk/test/flutter_google_places_sdk_test.dart
@@ -117,6 +117,8 @@ void main() {
                   "typeFilter": "ESTABLISHMENT",
                   'newSessionToken': false,
                   'origin': origin.toMap(),
+                  'locationBias': null,
+                  'locationRestriction': null,
                 }),
           ],
         );